+++ src/steps/access/converters.test.ts
@@ -1,4 +1,4 @@
-import { User } from '@jupiterone/tenable-client-nodejs';
+import { User } from '../../tenable/client';
 import { Entities } from '../../constants';
 import { Account } from '../../types';
 import { createAccountUserRelationship, createUserEntity } from './converters';