+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [8.5.0] 2022-06-08
+
 - move @jupiterone/tenable-client-nodejs to this repo
 - add assetMacAddress and agentId to vulnerability entity